double -> numeric

+++ conf/evolutions/default/1.sql
@@ -15,8 +15,8 @@ create table messages (
 create table rooms (
   id                        varchar(255) not null,
   name                      varchar(255),
-  latitude                  double precision,
-  longitude                 double precision,
+  latitude                  NUMERIC,
+  longitude                 NUMERIC,
   radius                    integer,
   creation_time             timestamp,
   last_activity             timestamp,
